+++ packages/clip/example/lib/main.dart
@@ -64,83 +64,78 @@ class _MyHomePageState extends State<MyHomePage> {
         child: SingleChildScrollView(
           child: Column(
             children: <Widget>[
-              ImageClipField(
-                key: ValueKey('avatar'),
-                initialValue: _user.avatar,
-                quality: 50,
-                maxHeight: 1024,
-                decoration: InputDecoration(
-                  helperText: 'Avatar',
-                ),
-                validator: (pickedFile) =>
-                    pickedFile == null ? 'Please upload attachment' : null,
-                builder: (context, pickedFile) {
-                  return CircleAvatar(
-                    radius: 56,
-                    child: pickedFile != null
-                        ? null
-                        : Center(
-                            child: Icon(
-                              Icons.add_a_photo_outlined,
-                              size: 28,
-                              color: Colors.white,
-                            ),
-                          ),
-                    backgroundImage: pickedFile != null
-                        ? FileImage(File(pickedFile.path))
-                        : null,
-                  );
-                },
-                onSaved: (avatar) {
-                  print(avatar);
-                },
-              ),
-              ListView(
-                shrinkWrap: true,
-                physics: NeverScrollableScrollPhysics(),
-                children: [
-                  ClipGridView(
-                    shrinkWrap: true,
-                    physics: const NeverScrollableScrollPhysics(),
-                    key: const ValueKey('avatars'),
-                    initialValues: <String?>[
-                      _user.avatar,
-                    ],
-                    quality: 20,
-                    maxHeight: 1024,
-                    decoration: InputDecoration(
-                      helperText: 'Please pick images',
-                    ),
-                    validator: (a) => a!.isEmpty ? 'aa' : null,
-                    autovalidateMode: AutovalidateMode.onUserInteraction,
-                    emptyBuilder: (context) {
-                      return CircleAvatar(
-                        radius: 56,
-                        child: Center(
-                          child: Icon(
-                            Icons.add_a_photo_outlined,
-                            size: 28,
-                            color: Colors.white,
-                          ),
-                        ),
-                      );
-                    },
-                    itemBuilder: (context, xFile, index) {
-                      return Card(
-                        child: CircleAvatar(
-                          radius: 56,
-                          backgroundImage: FileImage(File(xFile.path)),
-                        ),
-                      );
-                    },
-                    gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                      crossAxisCount: 3,
-                    ),
-                  ),
-                ],
-              ),
-              SizedBox(
-                height: 200,
+              // ImageClipField(
+              //   key: ValueKey('avatar'),
+              //   initialValue: _user.avatar,
+              //   quality: 50,
+              //   maxHeight: 1024,
+              //   decoration: InputDecoration(
+              //     helperText: 'Avatar',
+              //   ),
+              //   validator: (pickedFile) =>
+              //       pickedFile == null ? 'Please upload attachment' : null,
+              //   builder: (context, pickedFile) {
+              //     return CircleAvatar(
+              //       radius: 56,
+              //       child: pickedFile != null
+              //           ? null
+              //           : Center(
+              //               child: Icon(
+              //                 Icons.add_a_photo_outlined,
+              //                 size: 28,
+              //                 color: Colors.white,
+              //               ),
+              //             ),
+              //       backgroundImage: pickedFile != null
+              //           ? FileImage(File(pickedFile.path))
+              //           : null,
+              //     );
+              //   },
+              //   onSaved: (avatar) {
+              //     print(avatar);
+              //   },
+              // ),
+              // ClipGridView(
+              //   shrinkWrap: true,
+              //   physics: const NeverScrollableScrollPhysics(),
+              //   key: const ValueKey('avatars'),
+              //   initialValues: <String?>[
+              //     _user.avatar,
+              //   ],
+              //   quality: 20,
+              //   maxHeight: 1024,
+              //   decoration: InputDecoration(
+              //     helperText: 'Please pick images',
+              //   ),
+              //   validator: (a) => a!.isEmpty ? 'aa' : null,
+              //   autovalidateMode: AutovalidateMode.onUserInteraction,
+              //   emptyBuilder: (context) {
+              //     return CircleAvatar(
+              //       radius: 56,
+              //       child: Center(
+              //         child: Icon(
+              //           Icons.add_a_photo_outlined,
+              //           size: 28,
+              //           color: Colors.white,
+              //         ),
+              //       ),
+              //     );
+              //   },
+              //   itemBuilder: (context, xFile, index) {
+              //     return Card(
+              //       child: CircleAvatar(
+              //         radius: 56,
+              //         backgroundImage: FileImage(File(xFile.path)),
+              //       ),
+              //     );
+              //   },
+              //   gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+              //     crossAxisCount: 3,
+              //   ),
+              // ),
+              Container(
+                height: 126,
+                color: Colors.red,
                 child: ClipListView(
                   shrinkWrap: true,
                   separatorBuilder: (context, index) =>
@@ -152,28 +147,33 @@ class _MyHomePageState extends State<MyHomePage> {
                   ],
                   quality: 20,
                   maxHeight: 1024,
-                  decoration: InputDecoration(
-                    helperText: 'Please pick images',
-                  ),
-                  validator: (a) => a!.isEmpty ? 'aa' : null,
+                  validator: (a) => 'aa',
                   autovalidateMode: AutovalidateMode.onUserInteraction,
                   emptyBuilder: (context) {
-                    return CircleAvatar(
-                      radius: 56,
-                      child: Center(
-                        child: Icon(
-                          Icons.add_a_photo_outlined,
-                          size: 28,
-                          color: Colors.white,
+                    return SizedBox(
+                      height: 56,
+                      width: 56,
+                      child: CircleAvatar(
+                        radius: 56,
+                        child: Center(
+                          child: Icon(
+                            Icons.add_a_photo_outlined,
+                            size: 28,
+                            color: Colors.white,
+                          ),
                         ),
                       ),
                     );
                   },
                   itemBuilder: (context, xFile, index) {
-                    return Card(
-                      child: CircleAvatar(
-                        radius: 56,
-                        backgroundImage: FileImage(File(xFile.path)),
+                    return SizedBox(
+                      height: 56,
+                      width: 56,
+                      child: Card(
+                        child: CircleAvatar(
+                          radius: 56,
+                          backgroundImage: FileImage(File(xFile.path)),
+                        ),
                       ),
                     );
                   },